Package VirMet with setuptools plus minor fixes
`setup.py` was added such that the users can install the packages
with `python setup.py install`. More instructions for conda
might be added in the future. The structure was changed as
now `bin/virmet` only contains the import and call to
the main function in `virmet/__main__.py`. Without installing
the package, the user can call `path_to_dir/bin/virmet`to run it.
At the same time, the entry point defined in `setup.py` allows
the user to have it on their path after installation.

Since we are using `assing` pandas>=0.16 is required and added to
`environment.yml`.