More unit testing and installation instructions
Function `hunter` in `wolfpack.py` is tested now. This helped
fixing a few bugs that might have been critical on machines with
more than 100 CPUs. This test does not work on OS X because `seq -w 000 003`,
for example, returns 0 1 2 3 and not 000 001 002 003.

Installation instructions have been updated and we now try to install R
in Travis.